Improve mathematical expressions for clarity and precision
- Updated `result1` to use double precision in division.
- Refined `result11` for better order of operations and clarity.
- Split `result21` into two lines for improved readability.
- Removed unnecessary type casting in `result22`.

+++ C#2/Program.cs
@@ -4,7 +4,7 @@
     {
         static void Main(string[] args)
         {
-            int result1 = (3 + 5 * (6 + 3) - 8 * 3 + 1) / (21 + 50 / (3 + 4 * (1 + 2)));
+            double result1 = (3 + 5 * (6 + 3) - 8 * 3 + 1) / (21 + 50 /(double) (3 + 4 * (1 + 2)));
             double result2 = ((0.25 - 0.12) * 0.81 + 0.132 - 2.7 / 3.1) / (3 / 7 + 2.97 * (8.05 - 8.1 * 6.07));
             Console.WriteLine("Результаты вычислений:");
             Console.WriteLine(result1);
@@ -18,10 +18,10 @@
             Console.Write("Введите третье целое число:");
             int c = int.Parse(Console.ReadLine());
 
-            double result11 = (double)a + b - c * (3 * a * b + Math.Pow(a, 2) / (b * c) - Math.Pow((c + (a * b) / c), 2));
+            double result11 = a + b - c * (3 * a * b + Math.Pow(a, 2)) / (b * c) - Math.Pow(c + (double)(a * b) / c, 2);
 
-            double result21 = (double)Math.Pow(a * b + 7 * c, 3) - (Math.Pow(a, 2) + Math.Pow(b, 2) + Math.Pow(c, 2));
-            double result22 = result21 / (double)(b - a * c - a * (b + c)); /* Второй пример был разбит на 2 переменные 
+            double result21 = Math.Pow(a * b + 7 * c, 3) - (Math.Pow(a, 2) + Math.Pow(b, 2) + Math.Pow(c, 2));
+            double result22 = result21 / (b - a * c - a * (b + c)); /* Второй пример был разбит на 2 переменные 
             в целях удобства, поскольку весь пример не вмещался в одну строку*/
 
             Console.WriteLine($"Результат первого примера равен: {result11}");
